# proyecto3/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Banco XYZ Services
  description: API para gestionar transacciones del Banco XYZ
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Servidor local
paths:
...
  /transacciones/historial:
    get:
      tags:
        - transaccion
      summary: Listar transacciones
      operationId: listTransactions.
      responses:
        '200':
          description: Lista de transacciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionResponse'
        default:
          description: Error en la obtención de transacciones
  /transacciones/deposito:
    post:
      tags:
        - transaccion
      summary: Crear una transaccion de tipo Deposito
      operationId: registerTransactionDeposito
      requestBody:
        description: Registrar una transaccion de tipo Deposito
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionRequest"
      responses:
        '200':
          description: Transaccion tipo Deposito registrada exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        default:
          description: Error en la creacion de transaccion tipo Deposito
  /transacciones/retiro:
    post:
      tags:
        - transaccion
      summary: Crear una transaccion de tipo Retiro
      operationId: registerTransactionRetiro
      requestBody:
        description: Registrar una transaccion de tipo Retiro
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionRequest"
      responses:
        '200':
          description: Transaccion tipo Retiro registrada exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        default:
          description: Error en la creacion de transaccion tipo Retiro
  /transacciones/transferencia:
    post:
      tags:
        - transaccion
      summary: Crear una transaccion de tipo Transferencia
      operationId: registerTransactionTransferencia
      requestBody:
        description: Registrar una transaccion de tipo Transferencia
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionRequest"
      responses:
        '200':
          description: Transaccion tipo Transferencia registrada exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        default:
          description: Error en la creacion de transaccion tipo Transferencia
components:
  schemas:
...
    TransactionRequest:
      type: object
      properties:
        id:
          type: string
          description: ID único de la transaccion
        tipo:
          type: string
          description: Tipo de transaccion (DEPOSITO, RETIRO, TRANSFERENCIA)
        monto:
          type: number
          format: double
          description: Monto de la transaccion
        fecha:
          type: string
          format: date
          description: Fecha de la transaccion
        cuentaOrigen:
          type: string
          description: Cuenta Origen de la transaccion RETIRO o TRANSFERENCIA
        cuentaDestino:
          type: string
          description: Cuenta Destino de la transaccion DEPOSITO o TRANSFERENCIA
    TransactionResponse:
      type: object
      properties:
        id:
          type: string
          description: ID único de la transaccion
        tipo:
          type: string
          description: Tipo de transaccion (DEPOSITO, RETIRO, TRANSFERENCIA)
        monto:
          type: number
          format: double
          description: Monto de la transaccion
        fecha:
          type: string
          format: date
          description: Fecha de la transaccion
        cuentaOrigen:
          type: string
          description: Cuenta Origen de la transaccion RETIRO o TRANSFERENCIA
        cuentaDestino:
          type: string
          description: Cuenta Destino de la transaccion DEPOSITO o TRANSFERENCIA